chore: 增加对 uTools 环境的判断

+++ app/store/chat.ts
@@ -27,7 +27,7 @@ import { createPersistStore } from "../utils/store";
 import { collectModelsWithDefaultModel } from "../utils/model";
 import { useAccessStore } from "./access";
 import { isDalle3 } from "../utils";
-import { clearGlobalAsk, setupGlobalAsk } from "../utils/utools";
+import { clearGlobalAsk, isUTools, setupGlobalAsk } from "../utils/utools";
 
 export type ChatMessage = RequestMessage & {
   date: string;
@@ -670,11 +670,16 @@ export const useChatStore = createPersistStore(
       },
 
       clearAllData() {
-        utools.db.promises.allDocs().then((docs) => {
-          docs.forEach((doc) => {
-            utools.db.promises.remove(doc._id);
+        if (isUTools) {
+          utools.db.promises.allDocs().then((docs) => {
+            docs.forEach((doc) => {
+              utools.db.promises.remove(doc._id);
+            });
           });
-        });
+        } else {
+          // @ts-expect-error - Browser environment
+          storage.clear();
+        }
         location.reload();
       },
     };
